15/11/23 11:16 Guito Cambio de Long a Integer  id

+++ src/main/java/com/sisu/sisu/Dao/FarmaceuticaDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.FormaFarmaceutica;
 
-public interface FarmaceuticaDao extends CrudRepository<FormaFarmaceutica, Long>{
+public interface FarmaceuticaDao extends CrudRepository<FormaFarmaceutica, Integer>{
     
 }
