+++ src/main/java/com/sisu/sisu/Dao/HistoriaLinameDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.HistorialLiname;
 
-public interface HistoriaLinameDao extends CrudRepository<HistorialLiname, Long>{
+public interface HistoriaLinameDao extends CrudRepository<HistorialLiname, Integer>{
     
 }
