+++ src/main/java/com/sisu/sisu/Dao/IRemediosFarmaciaLoteDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.RemediosFarmaciaLote;
 
-public interface IRemediosFarmaciaLoteDao extends CrudRepository <RemediosFarmaciaLote, Long> {
+public interface IRemediosFarmaciaLoteDao extends CrudRepository <RemediosFarmaciaLote, Integer> {
     
 }
