+++ src/main/java/com/sisu/sisu/Service/ITipoProveedorService.java
@@ -7,6 +7,6 @@ import com.sisu.sisu.entitys.TipoProveedor;
 public interface ITipoProveedorService {
     public List<TipoProveedor> findAll();
     public void save(TipoProveedor tipoProveedor);
-    public TipoProveedor findOne(Long id);
-    public void delete(Long id);
+    public TipoProveedor findOne(Integer id);
+    public void delete(Integer id);
 }
